Update Dockerfile to include package.json files for dependency installation; modify ChatMessage interface to be exported for broader accessibility; fix income variable usage in AIService for correct income display.

+++ backend/src/services/ai/AIService.ts
@@ -250,7 +250,7 @@ Chỉ trả về JSON, không giải thích gì thêm.
     const prompt = `
 Bạn là chuyên gia tư vấn tài chính. Hãy đề xuất cách chia tiền vào các quỹ tiết kiệm.
 
-Thu nhập tháng: ${income.toLong()} VND
+Thu nhập tháng: ${income} VND
 Chi tiêu hiện tại:
 ${expensesJson}
